Reorganizes entire library
